+++ Tests/ZcashPaymentURITests/ZIP321Tests.swift
@@ -192,6 +192,92 @@ final class ZcashSwiftPaymentUriTests: XCTestCase {
                 )
             }
         }
+    }
+    
+    func testParsingMultiplePaymentsRequestStartingWithNoParamIndexAndNoAmount() throws {
+        let uriString = "zcash:?address=tmEZhbWHTpdKMw5it8YDspUXSMGQyFwovpU&address.1=ztestsapling10yy2ex5dcqkclhc7z7yrnjq2z6feyjad56ptwlfgmy77dmaqqrl9gyhprdx59qgmsnyfska2kez&amount.1=0.789&memo.1=VGhpcyBpcyBhIHVuaWNvZGUgbWVtbyDinKjwn6aE8J-PhvCfjok"
 
+        let address0 = "tmEZhbWHTpdKMw5it8YDspUXSMGQyFwovpU"
+
+        guard let recipient0 = RecipientAddress(value: address0, context: .testnet) else {
+            XCTFail("failed to create recipient without validation for address: \(address0)")
+            return
+        }
+
+        let payment0 = try Payment(
+            recipientAddress: recipient0,
+            amount: nil,
+            memo: nil,
+            label: nil,
+            message: nil,
+            otherParams: nil
+        )
+
+        let address1 = "ztestsapling10yy2ex5dcqkclhc7z7yrnjq2z6feyjad56ptwlfgmy77dmaqqrl9gyhprdx59qgmsnyfska2kez"
+
+        guard let recipient1 = RecipientAddress(value: address1, context: .testnet) else {
+            XCTFail("failed to create recipient without validation for address: \(address1)")
+            return
+        }
+
+        let payment1 = try Payment(
+            recipientAddress: recipient1,
+            amount: try Amount(value: 0.789),
+            memo: try MemoBytes(utf8String: "This is a unicode memo ✨🦄🏆🎉"),
+            label: nil,
+            message: nil,
+            otherParams: nil
+        )
+
+        let paymentRequest = try PaymentRequest(payments: [payment0, payment1])
+
+        let result = try ZIP321.request(from: uriString, context: .testnet)
+
+        XCTAssertNoDifference(result, ParserResult.request(paymentRequest))
+        
+        XCTAssertNoDifference(uriString, ZIP321.uriString(from: paymentRequest, formattingOptions: .useEmptyParamIndex(omitAddressLabel: false)))
+    }
+    
+    func testParsingMultiplePaymentsRequestStartingWithNoParamIndexIndexedParamHasNoAmount() throws {
+        let uriString = "zcash:?address=tmEZhbWHTpdKMw5it8YDspUXSMGQyFwovpU&amount=123.456&address.1=ztestsapling10yy2ex5dcqkclhc7z7yrnjq2z6feyjad56ptwlfgmy77dmaqqrl9gyhprdx59qgmsnyfska2kez&memo.1=VGhpcyBpcyBhIHVuaWNvZGUgbWVtbyDinKjwn6aE8J-PhvCfjok"
+
+        let address0 = "tmEZhbWHTpdKMw5it8YDspUXSMGQyFwovpU"
+
+        guard let recipient0 = RecipientAddress(value: address0, context: .testnet) else {
+            XCTFail("failed to create recipient without validation for address: \(address0)")
+            return
+        }
+
+        let payment0 = try Payment(
+            recipientAddress: recipient0,
+            amount: try Amount(value: 123.456),
+            memo: nil,
+            label: nil,
+            message: nil,
+            otherParams: nil
+        )
+
+        let address1 = "ztestsapling10yy2ex5dcqkclhc7z7yrnjq2z6feyjad56ptwlfgmy77dmaqqrl9gyhprdx59qgmsnyfska2kez"
+
+        guard let recipient1 = RecipientAddress(value: address1, context: .testnet) else {
+            XCTFail("failed to create recipient without validation for address: \(address1)")
+            return
+        }
+
+        let payment1 = try Payment(
+            recipientAddress: recipient1,
+            amount: nil,
+            memo: try MemoBytes(utf8String: "This is a unicode memo ✨🦄🏆🎉"),
+            label: nil,
+            message: nil,
+            otherParams: nil
+        )
+
+        let paymentRequest = try PaymentRequest(payments: [payment0, payment1])
+
+        let result = try ZIP321.request(from: uriString, context: .testnet)
+
+        XCTAssertNoDifference(result, ParserResult.request(paymentRequest))
+        XCTAssertNoDifference(uriString, ZIP321.uriString(from: paymentRequest, formattingOptions: .useEmptyParamIndex(omitAddressLabel: false)))
     }
 }